+++ frontend/src/metabase/timelines/containers/EditEventModal/EditEventModal.tsx
@@ -1,13 +1,13 @@
 import { connect } from "react-redux";
-import { goBack } from "react-router-redux";
+import { goBack, push } from "react-router-redux";
 import _ from "underscore";
 import * as Urls from "metabase/lib/urls";
 import Collections from "metabase/entities/collections";
 import Timelines from "metabase/entities/timelines";
 import TimelineEvents from "metabase/entities/timeline-events";
+import { Collection, Timeline, TimelineEvent } from "metabase-types/api";
 import { State } from "metabase-types/store";
 import EditEventModal from "../../components/EditEventModal";
-import { archiveEvent, updateEvent } from "../../actions";
 import { ModalProps } from "../../types";
 
 const timelineProps = {
@@ -26,11 +26,27 @@ const collectionProps = {
     Urls.extractCollectionId(props.params.slug),
 };
 
-const mapDispatchToProps = {
-  onSubmit: updateEvent,
-  onArchive: archiveEvent,
-  onCancel: goBack,
-};
+const mapDispatchToProps = (dispatch: any) => ({
+  onSubmit: async (
+    event: TimelineEvent,
+    timeline: Timeline,
+    collection: Collection,
+  ) => {
+    await dispatch(TimelineEvents.actions.update(event));
+    dispatch(push(Urls.timelineInCollection(timeline, collection)));
+  },
+  onArchive: async (
+    event: TimelineEvent,
+    timeline: Timeline,
+    collection: Collection,
+  ) => {
+    await dispatch(TimelineEvents.actions.setArchived(event, true));
+    dispatch(push(Urls.timelineInCollection(timeline, collection)));
+  },
+  onCancel: () => {
+    dispatch(goBack());
+  },
+});
 
 export default _.compose(
   Timelines.load(timelineProps),
