+++ frontend/src/metabase/timelines/containers/NewTimelineModal/NewTimelineModal.tsx
@@ -1,11 +1,12 @@
 import { connect } from "react-redux";
-import { goBack } from "react-router-redux";
+import { goBack, push } from "react-router-redux";
 import _ from "underscore";
 import * as Urls from "metabase/lib/urls";
 import Collections from "metabase/entities/collections";
+import Timelines from "metabase/entities/timelines";
+import { Collection, Timeline } from "metabase-types/api";
 import { State } from "metabase-types/store";
 import NewTimelineModal from "../../components/NewTimelineModal";
-import { createTimeline } from "../../actions";
 import { ModalProps } from "../../types";
 
 const collectionProps = {
@@ -13,10 +14,17 @@ const collectionProps = {
     Urls.extractCollectionId(props.params.slug),
 };
 
-const mapDispatchToProps = {
-  onSubmit: createTimeline,
-  onCancel: goBack,
-};
+const mapDispatchToProps = (dispatch: any) => ({
+  onSubmit: async (values: Partial<Timeline>, collection: Collection) => {
+    const action = Timelines.actions.create(values);
+    const response = await dispatch(action);
+    const timeline = Timelines.HACK_getObjectFromAction(response);
+    dispatch(push(Urls.timelineInCollection(timeline, collection)));
+  },
+  onCancel: () => {
+    dispatch(goBack());
+  },
+});
 
 export default _.compose(
   Collections.load(collectionProps),
