+++ frontend/src/metabase/timelines/containers/TimelineModal/TimelineModal.tsx
@@ -3,10 +3,11 @@ import _ from "underscore";
 import * as Urls from "metabase/lib/urls";
 import Collections from "metabase/entities/collections";
 import Timelines from "metabase/entities/timelines";
+import TimelineEvents from "metabase/entities/timeline-events";
+import { TimelineEvent } from "metabase-types/api";
 import { State } from "metabase-types/store";
 import TimelineModal from "../../components/TimelineModal";
 import { ModalProps } from "../../types";
-import { archiveEvent } from "metabase/timelines/actions";
 
 const timelineProps = {
   id: (state: State, props: ModalProps) =>
@@ -19,9 +20,11 @@ const collectionProps = {
     Urls.extractCollectionId(props.params.slug),
 };
 
-const mapDispatchToProps = {
-  onArchive: archiveEvent,
-};
+const mapDispatchToProps = (dispatch: any) => ({
+  onArchive: async (event: TimelineEvent) => {
+    await dispatch(TimelineEvents.actions.setArchived(event, true));
+  },
+});
 
 export default _.compose(
   Timelines.load(timelineProps),
